refactor(cl-agg): use zkvm abstraction in guest
