+++ crates/proof-impl/cl-agg/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "strata-proofimpl-cl-agg"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+borsh.workspace = true
+strata-zkvm.workspace = true
+strata-proofimpl-cl-stf.workspace = true
