+++ provers/sp1/guest-cl-agg/Cargo.toml
@@ -6,13 +6,8 @@ version = "0.1.0"
 [workspace]
 
 [dependencies]
-bincode = "1.3.3"
-borsh = { version = "1.5.0", features = ["derive"] }
-sha2 = "0.10.8"
-sp1-zkvm = { version = "3.0.0", features = ["verify"] }
-strata-primitives = { path = "../../../crates/primitives" }
-strata-proofimpl-checkpoint = { path = "../../../crates/proof-impl/checkpoint" }
-strata-proofimpl-cl-stf = { path = "../../../crates/proof-impl/cl-stf" }
+strata-proofimpl-cl-agg = { path = "../../../crates/proof-impl/cl-agg" }
+strata-sp1-adapter = { path = "../../../crates/zkvm/adapters/sp1" }
 
 [patch.crates-io]
 secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", branch = "patch-secp256k1-v0.29.1" }
